Add Test/Functional/WebTestCase; deprecated Test/WebTestCase; setUp and tearDown downgraded to protected; coding style cleanup
