+++ Test/DependencyInjection/ExtensionTestCase.php
@@ -25,7 +25,7 @@ abstract class ExtensionTestCase extends TestCase
     /**
      * {@inheritdoc}
      */
-    public function setUp()
+    protected function setUp()
     {
         parent::setUp();
 
@@ -35,7 +35,7 @@ abstract class ExtensionTestCase extends TestCase
     /**
      * {@inheritdoc}
      */
-    public function tearDown()
+    protected function tearDown()
     {
         unset($this->container);
