+++ Test/Helper/ValidatorHelper.php
@@ -7,8 +7,7 @@ namespace IC\Bundle\Base\TestBundle\Test\Helper;
 
 use Symfony\Component\Validator\ConstraintValidator;
 use Symfony\Component\Validator\Constraint;
-
-use IC\Bundle\Base\TestBundle\Test\WebTestCase;
+use IC\Bundle\Base\TestBundle\Test\Functional\WebTestCase;
 
 /**
  * Validator helper class.
